Added the 3 sequences and their tests

+++ src/magic_squares.py
@@ -1,2 +1,3 @@
 def magic_square(n):
-    return n*((n^2)+1)/2
+    return int(n*((n^2)+1)/2)
+
